+++ wp-video-grid.php
@@ -3,7 +3,7 @@
  * Plugin Name: WordPress Video Grid
  * Plugin URI: https://ajithrn.com/wp-video-grid
  * Description: Plugin to manage and display videos in grid
- * Version: 1.0.0
+ * Version: 1.2.0
  * Author: Ajith
  * Author URI: https://ajithrn.com
  * License: GPL-2.0+
@@ -18,7 +18,7 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 // Define plugin constants
-define( 'WP_VIDEO_GRID_VERSION', '1.0.0' );
+define( 'WP_VIDEO_GRID_VERSION', '1.2.0' );
 define( 'WP_VIDEO_GRID_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'WP_VIDEO_GRID_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 
@@ -30,6 +30,7 @@ require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Block.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Frontend.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Admin.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Ajax.php';
+require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Settings.php';
 
 /**
  * Initialize the plugin
@@ -42,6 +43,7 @@ function wp_video_grid_init() {
     new WP_Video_Grid\Frontend();
     new WP_Video_Grid\Admin();
     new WP_Video_Grid\Ajax();
+    new WP_Video_Grid\Settings();
 }
 add_action( 'plugins_loaded', 'wp_video_grid_init' );
 
@@ -59,7 +61,9 @@ add_action( 'admin_enqueue_scripts', 'wp_video_grid_admin_scripts' );
  * Activation hook
  */
 function wp_video_grid_activate() {
-    // Activation code here
+    
+    // Add default option for data deletion
+    add_option('wp_video_grid_delete_data', 'no');
     flush_rewrite_rules();
 }
 register_activation_hook( __FILE__, 'wp_video_grid_activate' );